requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.1
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
diskcache==5.6.3
exceptiongroup==1.1.3
executing==2.0.0
FLAML==2.1.1
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.4.0
html5lib==1.1
idna==3.4
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
multidict==6.0.4
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.1
openai==0.28.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyautogen==0.1.13
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
termcolor==2.3.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
webencodings==0.5.1
yarl==1.9.2
yfinance==0.2.31
